requirements.txt
pillow
inflect
random
numpy
pandas
matplotlib.pyplot
tensorflow
io
flask
sklearn